+++ src/components/LangSwitch.vue
@@ -1,7 +1,9 @@
 <script setup lang="ts">
-import { computed } from "vue";
+import { computed, onMounted } from "vue";
 import "@/../node_modules/flag-icons/css/flag-icons.min.css";
 import store from "@/store/sessionData";
+import i18n from "@/i18n";
+
 const locales = [
   {
     name: "Polski",
@@ -23,14 +25,18 @@ const currentLocale = computed(() => store.state.locale);
 </script>
 
 <template>
-  <div @load="$i18n.locale = currentLocale" class="flex space-x-2 pt-1.5">
+  <div class="flex space-x-2 pt-1.5" @loadstart="$i18n.locale = currentLocale">
     <button
       v-for="locale in locales"
       :key="locale.lang"
       @click="
         changeLocale(locale.lang);
-        console.log(`${locale.lang}`);
         $i18n.locale = locale.lang;
+      "
+      @load="
+        console.log($i18n.locale);
+        console.log(currentLocale);
+        console.log(locale.lang);
       "
       :class="[currentLocale === locale.lang ? 'opacity-100' : 'opacity-30']"
       :disabled="currentLocale === locale.lang"
