+++ src/cookies/setCookie.ts
@@ -1,0 +1,6 @@
+// Helper function to set a cookie
+export function setCookie(name: string, value: string, days: number) {
+  const expires = new Date();
+  expires.setTime(expires.getTime() + days * 24 * 60 * 60 * 1000);
+  document.cookie = `${name}=${value};expires=${expires.toUTCString()};path=/`;
+}
