+++ src/locales/en.json
@@ -1,3 +1,5 @@
 {
-    "hello": "hello"
+    "hello": "hello",
+    "main": "main",
+    "map": "map"
 }
