+++ src/locales/pl.json
@@ -1,3 +1,5 @@
 {
-    "hello": "witaj"
+    "hello": "witaj",
+    "main": "główna",
+    "map": "mapa"
 }
