+++ src/main.ts
@@ -5,11 +5,11 @@ import store from "./store";
 import "./index.css";
 import { messages, defaultLocale } from "@/locales";
 import { createI18n } from "vue-i18n";
+import { getCookie } from "./cookies/getCookie";
 
 const i18n = createI18n({
   messages,
-  locale: defaultLocale,
-  fallbackLocale: defaultLocale,
+  locale: getCookie("locale") || defaultLocale,
 });
 
 createApp(App).use(store).use(router).use(i18n).mount("#app");
