+++ src/store/sessionData.ts
@@ -1,24 +1,7 @@
 import { createStore } from "vuex";
 import { defaultLocale } from "@/locales";
-
-// Helper function to set a cookie
-function setCookie(name: string, value: string, days: number) {
-  const expires = new Date();
-  expires.setTime(expires.getTime() + days * 24 * 60 * 60 * 1000);
-  document.cookie = `${name}=${value};expires=${expires.toUTCString()};path=/`;
-}
-
-// Helper function to get a cookie
-function getCookie(name: string | string[]) {
-  const cookies = document.cookie.split(";");
-  for (let i = 0; i < cookies.length; i++) {
-    const cookie = cookies[i].trim();
-    if (cookie.startsWith(name + "=")) {
-      return cookie.substring(name.length + 1);
-    }
-  }
-  return null;
-}
+import { getCookie } from "../cookies/getCookie";
+import { setCookie } from "../cookies/setCookie";
 
 // Define your Vuex store here
 const store = createStore({
